C: Rapikan jawaban
