+++ Faktor/C/faktor.c
